requirements.txt
requests
dnspython
colorama
rich